requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.10.0
transformers>=4.15.0
sentence-transformers>=2.2.0
flask>=2.0.0
faiss-cpu>=1.7.0
jupyter>=1.0.0
matplotlib>=3.4.0
tqdm>=4.62.0
python-dotenv>=0.19.0
pytest>=6.2.5
